+++ src/main/java/com/alon/spring/crud/service/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package com.alon.spring.crud.service;
+package com.alon.spring.crud.service.exception;
 
-public class NotFoundException extends Exception {
+public class NotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;